replace old command-based template
