--- wpilibcExamples/src/main/cpp/templates/commandbased/cpp/OI.cpp
@@ -1,12 +1,0 @@
-/*----------------------------------------------------------------------------*/
-/* Copyright (c) 2017-2019 FIRST. All Rights Reserved.                        */
-/* Open Source Software - may be modified and shared by FRC teams. The code   */
-/* must be accompanied by the FIRST BSD license file in the root directory of */
-/* the project.                                                               */
-/*----------------------------------------------------------------------------*/
-
-#include "OI.h"
-
-OI::OI() {
-  // Process operator interface input here.
-}
