+++ wpilibcExamples/src/main/cpp/templates/commandbased/include/Constants.h
