+++ wpilibcExamples/src/main/cpp/templates/commandbased/include/commands/ExampleCommand.h
@@ -1,5 +1,5 @@
 /*----------------------------------------------------------------------------*/
-/* Copyright (c) 2017-2018 FIRST. All Rights Reserved.                        */
+/* Copyright (c) 2019 FIRST. All Rights Reserved.                             */
 /* Open Source Software - may be modified and shared by FRC teams. The code   */
 /* must be accompanied by the FIRST BSD license file in the root directory of */
 /* the project.                                                               */
@@ -7,14 +7,28 @@
 
 #pragma once
 
-#include <frc/commands/Command.h>
+#include <frc2/command/CommandHelper.h>
+#include <frc2/command/SendableCommandBase.h>
 
-class ExampleCommand : public frc::Command {
+#include "subsystems/ExampleSubsystem.h"
+
+/**
+ * An example command that uses an example subsystem.
+ *
+ * <p>Note that this extends CommandHelper, rather extending SendableCommandBase
+ * directly; this is crucially important, or else the decorator functions in
+ * Command will *not* work!
+ */
+class ExampleCommand
+    : public frc2::CommandHelper<frc2::SendableCommandBase, ExampleCommand> {
  public:
-  ExampleCommand();
-  void Initialize() override;
-  void Execute() override;
-  bool IsFinished() override;
-  void End() override;
-  void Interrupted() override;
+  /**
+   * Creates a new ExampleCommand.
+   *
+   * @param subsystem The subsystem used by this command.
+   */
+  explicit ExampleCommand(ExampleSubsystem* subsystem);
+
+ private:
+  ExampleSubsystem* m_subsystem;
 };
