+++ wpilibcExamples/src/main/cpp/templates/commandbased/include/subsystems/ExampleSubsystem.h
@@ -1,5 +1,5 @@
 /*----------------------------------------------------------------------------*/
-/* Copyright (c) 2017-2018 FIRST. All Rights Reserved.                        */
+/* Copyright (c) 2019 FIRST. All Rights Reserved.                             */
 /* Open Source Software - may be modified and shared by FRC teams. The code   */
 /* must be accompanied by the FIRST BSD license file in the root directory of */
 /* the project.                                                               */
@@ -7,14 +7,20 @@
 
 #pragma once
 
-#include <frc/commands/Subsystem.h>
+#include <frc2/command/SendableSubsystemBase.h>
 
-class ExampleSubsystem : public frc::Subsystem {
+class ExampleSubsystem : public frc2::SendableSubsystemBase {
  public:
   ExampleSubsystem();
-  void InitDefaultCommand() override;
+
+  /**
+   * Will be called periodically whenever the CommandScheduler runs.
+   */
+  void Periodic() override;
+
+  // Subsystem methods go here.
 
  private:
-  // It's desirable that everything possible under private except
-  // for methods that implement subsystem capabilities
+  // Components (e.g. motor controllers and sensors) should generally be
+  // declared private and exposed only through public methods.
 };
