--- wpilibcExamples/src/main/cpp/templates/commandbasednew/cpp/subsystems/ExampleSubsystem.cpp
@@ -1,16 +1,0 @@
-/*----------------------------------------------------------------------------*/
-/* Copyright (c) 2019 FIRST. All Rights Reserved.                             */
-/* Open Source Software - may be modified and shared by FRC teams. The code   */
-/* must be accompanied by the FIRST BSD license file in the root directory of */
-/* the project.                                                               */
-/*----------------------------------------------------------------------------*/
-
-#include "subsystems/ExampleSubsystem.h"
-
-ExampleSubsystem::ExampleSubsystem() {
-  // Implementation of subsystem constructor goes here.
-}
-
-void ExampleSubsystem::Periodic() {
-  // Implementation of subsystem periodic method goes here.
-}
