--- wpilibcExamples/src/main/cpp/templates/commandbasednew/include/commands/ExampleCommand.h
@@ -1,34 +1,0 @@
-/*----------------------------------------------------------------------------*/
-/* Copyright (c) 2019 FIRST. All Rights Reserved.                             */
-/* Open Source Software - may be modified and shared by FRC teams. The code   */
-/* must be accompanied by the FIRST BSD license file in the root directory of */
-/* the project.                                                               */
-/*----------------------------------------------------------------------------*/
-
-#pragma once
-
-#include <frc2/command/CommandHelper.h>
-#include <frc2/command/SendableCommandBase.h>
-
-#include "subsystems/ExampleSubsystem.h"
-
-/**
- * An example command that uses an example subsystem.
- *
- * <p>Note that this extends CommandHelper, rather extending SendableCommandBase
- * directly; this is crucially important, or else the decorator functions in
- * Command will *not* work!
- */
-class ExampleCommand
-    : public frc2::CommandHelper<frc2::SendableCommandBase, ExampleCommand> {
- public:
-  /**
-   * Creates a new ExampleCommand.
-   *
-   * @param subsystem The subsystem used by this command.
-   */
-  explicit ExampleCommand(ExampleSubsystem* subsystem);
-
- private:
-  ExampleSubsystem* m_subsystem;
-};
