+++ wpilibcExamples/src/main/cpp/templates/templates.json
@@ -43,14 +43,5 @@
     ],
     "foldername": "commandbased",
     "gradlebase": "cpp"
-  },
-  {
-    "name": "Command Robot (New)",
-    "description": "Command style",
-    "tags": [
-      "Command"
-    ],
-    "foldername": "commandbasednew",
-    "gradlebase": "cpp"
   }
 ]
